--HG--
branch : server-fixes

+++ Automation/psi_update_stats_dat.py
@@ -21,14 +21,20 @@ import os
 import shlex
 import subprocess
 import psi_ops_stats_credentials
+import time
+import logging
 
 PSI_OPS_DB_FILENAME = os.path.join(os.path.abspath('.'), 'psi_ops_stats.dat')
 # Can't overwrite target file directly due to Wine limitation
 export_filename = os.path.join(os.path.abspath('.'), 'psi_ops_stats.dat.temp')
 
-if __name__ == "__main__":
+MAX_RETRIES = 3
+RETRY_SLEEP_TIME = 150
+
+def update_stats_dat():
     try:
         if os.path.isfile(export_filename):
+            logging.info('Removing temp file')
             os.remove(export_filename)
         cmd = 'wine ./CipherShare/CipherShareScriptingClient.exe \
                 ExportDocument \
@@ -49,11 +55,35 @@ if __name__ == "__main__":
         output = proc.communicate()
 
         if proc.returncode != 0:
-            raise Exception('CipherShare export failed: ' + str(output))
+            msg = 'CipherShare export failed %s' % str(output)
+            logging.warning(msg)
+            raise Exception(msg)
 
-        if os.path.isfile(PSI_OPS_DB_FILENAME):
-            os.remove(PSI_OPS_DB_FILENAME)
-        os.rename(export_filename, PSI_OPS_DB_FILENAME)
+        if os.path.isfile(export_filename):
+            if os.path.isfile(PSI_OPS_DB_FILENAME):
+                os.remove(PSI_OPS_DB_FILENAME)
+                logging.info('%s removed', PSI_OPS_DB_FILENAME)
+            os.rename(export_filename, PSI_OPS_DB_FILENAME)
+            logging.info('%s renamed to %s', export_filename, PSI_OPS_DB_FILENAME)
+        else:
+            logging.info('%s is not found', export_filename)
+            logging.info('%s, %s', proc.stdout, proc.stderr)
 
     except Exception as e:
+        logging.warning('Exception: %s', str(e)) 
         print str(e)
+
+if __name__ == "__main__":
+    count = 0
+    stats_db_ctime = None
+    if os.path.exists(PSI_OPS_DB_FILENAME):
+        stats_db_ctime = os.path.getctime(PSI_OPS_DB_FILENAME)
+    logging.basicConfig(filename='psi_update_stats_dat.log', format='%(asctime)s %(levelname)s %(message)s', level=logging.DEBUG)
+    while count < MAX_RETRIES:
+        update_stats_dat()
+        if os.path.getctime(PSI_OPS_DB_FILENAME) > stats_db_ctime: 
+            count = MAX_RETRIES
+        else:
+            count += 1
+            logging.info('did not update')
+            time.sleep(RETRY_SLEEP_TIME)
